genai-rag-web-ingestion/ingestion/requirement.txt
requests
beautifulsoup4
azure-storage-blob
python-dotenv